# docs/swagger.json
{
  "components": {
    "schemas": {
...
      "accounts.getDelegates": {
        "id": "accounts.getDelegates",
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "format": "address",
            "minLength": 1,
            "maxLength": 22
          }
        },
        "required": [
          "address"
        ]
      },
      "accounts.getDelegatesFee": {
        "id": "accounts.getDelegatesFee",
        "type": "object",
        "properties": {
          "height": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "accounts.getAccount": {
        "id": "accounts.getAccount",
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "format": "address",
            "minLength": 1,
            "maxLength": 22
          },
          "publicKey": {
            "type": "string",
            "format": "publicKey"
          }
        }
      },
...
      "blocks.getFees": {
        "id": "blocks.getFees",
        "type": "object",
        "properties": {
          "height": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "delegates.getFee": {
        "id": "delegates.getFee",
        "type": "object",
        "properties": {
          "height": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "delegates.getForgedByAccount": {
        "id": "delegates.getForgedByAccount",
        "type": "object",
        "properties": {
          "generatorPublicKey": {
            "type": "string",
            "format": "publicKey"
          },
          "start": {
            "type": "integer"
          },
          "end": {
            "type": "integer"
          }
        },
        "required": [
          "generatorPublicKey"
        ]
      },
...
      "responses.accounts.getAccount": {
        "id": "responses.accounts.getAccount",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "account": {
            "type": "object",
            "properties": {
              "address": {
                "id": "common.scalars.address",
                "type": "string",
                "format": "address",
                "minLength": 1,
                "maxLength": 22
              },
              "balance": {
                "id": "common.scalars.balance",
                "type": "string"
              },
              "multisignatures": {
                "type": "array",
                "items": {
                  "id": "common.scalars.signature",
                  "type": "string",
                  "format": "binary"
                }
              },
              "publicKey": {
                "id": "common.scalars.publicKey",
                "type": "string",
                "format": "publicKey"
              },
              "secondPublicKey": {
                "id": "common.scalars.publicKey",
                "type": "string",
                "format": "publicKey"
              },
              "secondSignature": {
                "id": "common.scalars.boolInt",
                "type": "integer",
                "enum": [
                  1,
                  0
                ]
              },
              "u_multisignatures": {
                "type": "array",
                "items": {
                  "id": "common.scalars.signature",
                  "type": "string",
                  "format": "binary"
                }
              },
              "unconfirmedBalance": {
                "id": "common.scalars.balance",
                "type": "string"
              },
              "unconfirmedSignature": {
                "id": "common.scalars.boolInt",
                "type": "integer",
                "enum": [
                  1,
                  0
                ]
              }
            }
          }
        },
        "example": {
          "success": true,
          "account": {
            "address": "8093718274007724701R",
            "balance": "2973803650603",
            "multisignatures": [],
            "publicKey": "7067a911f3a4e13facbae9006b52a0c3ac9824bdd9f37168303152ae49dcb1c0",
            "secondPublicKey": "e26988a52c519c9766d6f32ec32202b1ab16e77f6e404134222552fb3df23565",
            "unconfirmedBalance": "2973803650603",
            "unconfirmedSignature": 1,
            "secondSignature": 1,
            "u_multisignatures": []
          }
        }
      },
...
      "responses.accounts.getDelegatesFee": {
        "id": "responses.accounts.getDelegatesFee",
        "type": "object",
        "proprties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "fee": {
            "type": "number"
          }
        },
        "example": {
          "success": true,
          "fee": 2500000000
        }
      },
...
      "responses.general.deprecated": {
        "id": "responses.general.deprecated",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "success": false,
          "error": "Method is deprecated"
        }
      },
...
    }
  },
  "info": {
    "title": "RISE Node",
    "version": "1.2.0",
...
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/accounts/": {
      "get": {
        "operationId": "AccountsAPI.getAccount",
        "parameters": [
          {
            "in": "query",
            "name": "accounts.getAccount",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/accounts.getAccount"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.accounts.getAccount"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get Account",
        "tags": [
          "Accounts API"
        ],
        "description": "A really nice **Description** for Get Account",
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.accounts\n  .getAccount('8093718274007724701R')\n  .then(function({ account }) {\n    console.log(account.balance) // 2973803650603\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
...
    "/api/accounts/delegates": {
      "get": {
        "operationId": "AccountsAPI.getDelegates",
        "parameters": [
          {
            "in": "query",
            "name": "accounts.getDelegates",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/accounts.getDelegates"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.accounts.getDelegates"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get delegates",
        "tags": [
          "Accounts API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.accounts\n  .getDelegates('8093718274007724701R')\n  .then(function({ delegates }) {\n    console.log(delegates[0].publicKey) // \"5d3c3c5cdead6...\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      },
      "put": {
        "operationId": "AccountsAPI.addDelegate",
        "responses": {
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.general.deprecated"
                }
              }
            },
            "description": "Deprecated API Error"
          }
        },
        "summary": "Add delegate",
        "tags": [
          "Accounts API"
        ],
        "deprecated": true
      }
    },
    "/api/accounts/delegates/fee": {
      "get": {
        "operationId": "AccountsAPI.getDelegatesFee",
        "parameters": [
          {
            "in": "query",
            "name": "accounts.getDelegatesFee",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/accounts.getDelegatesFee"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.accounts.getDelegatesFee"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get delegates fee",
        "tags": [
          "Accounts API"
        ]
      }
    },
...
    "/api/blocks/getFees": {
      "get": {
        "operationId": "BlocksAPI.getFees",
        "parameters": [
          {
            "in": "query",
            "name": "blocks.getFees",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/blocks.getFees"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.blocks.getFees"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get fees",
        "tags": [
          "Blocks API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.blocks\n  .getFeeSchedule()\n  .then(function({ fees }) {\n    console.log(fees.send) // 10000000\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
...
    "/api/delegates/": {
      "get": {
        "operationId": "DelegatesAPI.getDelegates",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.getDelegates",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.getDelegates"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getDelegates"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get delegates",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .getList()\n  .then(function({ delegates }) {\n    console.log(delegates[0].username) // official_pool\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      },
      "put": {
        "operationId": "DelegatesAPI.createDelegate",
        "responses": {
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.general.deprecated"
                }
              }
            },
            "description": "Deprecated API Error"
          }
        },
        "summary": "Create delegate",
        "tags": [
          "Delegates API"
        ],
        "deprecated": true
      }
    },
    "/api/delegates/fee": {
      "get": {
        "operationId": "DelegatesAPI.getFee",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.getFee",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.getFee"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getFee"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get fee",
        "tags": [
          "Delegates API"
        ]
      }
    },
    "/api/delegates/forging/getForgedByAccount": {
      "get": {
        "operationId": "DelegatesAPI.getForgedByAccount",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.getForgedByAccount",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.getForgedByAccount"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getForgedByAccount"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get forged by account",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .getForgedByAccount(\n    '7067a911f3a4e13facbae9006b52a0c3ac9824bdd9f37168303152ae49dcb1c0'\n  )\n  .then(function({ forged }) {\n    console.log(forged) // 19066025346961\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
    "/api/delegates/get": {
      "get": {
        "operationId": "DelegatesAPI.getDelegate",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.getDelegate",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.getDelegate"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getDelegate"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get delegate",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .getByPublicKey(\n    '7067a911f3a4e13facbae9006b52a0c3ac9824bdd9f37168303152ae49dcb1c0'\n  )\n  .then(function({ delegate }) {\n    console.log(delegate.username) // official_pool\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
    "/api/delegates/voters": {
      "get": {
        "operationId": "DelegatesAPI.getVoters",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.getVoters",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.getVoters"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getVoters"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get voters",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .getVoters('a65a8160b1e0733f66d1f1a8f322c9af29b26d5e491a84d6e3ae0ec43e000446')\n  .then(function({ accounts }) {\n    console.log(accounts.length) // 5\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
    "/api/delegates/search": {
      "get": {
        "operationId": "DelegatesAPI.search",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.search",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.search"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Search",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .search({ q: 'pool' })\n  .then(function({ delegates }) {\n    console.log(delegates[0].username) // official_pool\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
...
    "/api/delegates/getNextForgers": {
      "get": {
        "operationId": "DelegatesAPI.getNextForgers",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getNextForgers"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Get next forgers",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .getNextForgers()\n  .then(function({ delegates }) {\n    console.log(delegates[0]) // a65a8160b1e07...\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
    "/api/delegates/forging/status": {
      "get": {
        "operationId": "DelegatesAPI.getForgingStatus",
        "parameters": [
          {
            "in": "query",
            "name": "delegates.forgingStatus",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/delegates.forgingStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.getForgingStatus"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.accessDenied"
                }
              }
            },
            "description": "Delegates API Access Denied"
          }
        },
        "summary": "Get forging status",
        "tags": [
          "Delegates API"
        ],
        "x-code-samples": [
          {
            "lang": "javascript",
            "source": "import { rise } from 'risejs'\n\nrise.delegates\n  .getForgingStatus()\n  .then(function({ enabled }) {\n    console.log(enabled) // true\n  })\n  .catch(function(err) {\n    console.log('Error: ', err) // handle error\n  })\n"
          }
        ]
      }
    },
    "/api/delegates/forging/enable": {
      "post": {
        "operationId": "DelegatesAPI.forgingEnable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/delegates.disableForging"
              }
            }
          },
          "description": "delegates.disableForging",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.forgingEnable"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.accessDenied"
                }
              }
            },
            "description": "Delegates API Access Denied"
          }
        },
        "summary": "Forging enable",
        "tags": [
          "Delegates API"
        ]
      }
    },
    "/api/delegates/forging/disable": {
      "post": {
        "operationId": "DelegatesAPI.forgingDisable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/delegates.disableForging"
              }
            }
          },
          "description": "delegates.disableForging",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.forgingDisable"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responses.delegates.accessDenied"
                }
              }
            },
            "description": "Delegates API Access Denied"
          }
        },
        "summary": "Forging disable",
        "tags": [
          "Delegates API"
        ]
      }
    },
...
    "/api/multisignatures/accounts": {
      "get": {
        "operationId": "MultisignatureAPI.getAccounts",
        "parameters": [
          {
            "in": "query",
            "name": "multisignatures.getAccounts",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/multisignatures.getAccounts"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get accounts",
        "tags": [
          "Multisignature API"
        ]
      }
    },
...
  }
}